+++ pentagon/component/core/__init__.py
@@ -1,0 +1,5 @@
+from pentagon.component import ComponentBase
+
+
+class Core(ComponentBase):
+    pass