+++ pentagon/helpers.py
@@ -3,6 +3,8 @@ import os
 import traceback
 import jinja2
 import yaml
+from Crypto.PublicKey import RSA
+from stat import *
 
 
 def render_template(template_name, template_path, target, context, delete_template=True, overwrite=False):
@@ -23,22 +25,27 @@ def render_template(template_name, template_path, target, context, delete_templa
 
         logging.debug("Attempting to write {} from template {}{}".format(target, template_path, template_name))
 
+        template_path = os.path.normpath(template_path)
+        template_name = os.path.normpath(template_name)
+        template_permissions = os.stat(template_path + '/' + template_name).st_mode
+
         with open(target, 'w+') as vars_file:
             try:
-                template = jinja2.Environment(loader=jinja2.FileSystemLoader(os.path.normpath(template_path))).get_template(os.path.normpath(template_name))
+                template = jinja2.Environment(loader=jinja2.FileSystemLoader(template_path)).get_template(template_name)
                 vars_file.write(template.render(context))
             except Exception, e:
                 logging.error("Error writing {}. {}".format(target, traceback.print_exc(e)))
                 return False
+
+        os.chmod(target, template_permissions)
 
         if delete_template:
             logging.debug("Removing {}/{}".format(template_path, template_name))
             os.remove("{}/{}".format(template_path, template_name))
 
 
-def write_yaml_file(filename, dict):
+def write_yaml_file(filename, d):
     """ Accepts  filepath,  dictionary. Writes dictionary in yaml to file path, recursively creating path if necessary """
-
     if not os.path.exists(os.path.dirname(filename)):
         try:
             os.makedirs(os.path.dirname(filename))
@@ -47,4 +54,21 @@ def write_yaml_file(filename, dict):
                 raise
 
     with open(filename, "w") as f:
-        f.write(yaml.safe_dump(dict))
+        f.write(yaml.safe_dump(d, default_flow_style=False))
+
+
+def create_rsa_key(name, path, bits=2048):
+    """ creates an ssh key pair. Accepts name, path and bits. Name is the name of the key pair to generate at Path. Bits defaults to 2048 """
+
+    key = RSA.generate(bits)
+
+    private_key = "{}{}".format(path, name)
+    public_key = "{}{}.pub".format(path, name)
+
+    with open(private_key, 'w') as content_file:
+        os.chmod(private_key, 0600)
+        content_file.write(key.exportKey('PEM'))
+
+    pubkey = key.publickey()
+    with open(public_key, 'w') as content_file:
+        content_file.write(pubkey.exportKey('OpenSSH'))