+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Environment variables are now checked in ComponentBase class
 - Defaults to component
 - overwrite to template rendering
+- added inventory component
+- added vpn component
 
 ### Removed
 - env-vars.sh script
@@ -17,6 +19,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Updated 
 - makefile to support `yaml_source` change
+- added distutil.dir_util to allow overwriting exisint directories
+- added exit on failure for ComponentBase class
+- added default config out file for Pentaong start-project
+- updated config file output to sanitize and not include blank values
 
 ## [1.2.0] - 2017-11-8
 ### Added